@@ -1,3 +1,3 @@
 module github.com/japb1998/lashroom/cliApp
 
-go 1.20
+go 1.21
