@@ -7,19 +7,20 @@ import (
 )
 
 type ClientItem struct {
-	PrimaryKey   string    `json:"primaryKey"`
-	SortKey      string    `json:"sortKey"`
-	Phone        string    `json:"phone"`
-	Email        string    `json:"email"`
-	FirstName    string    `json:"firstName"`
-	LastName     string    `json:"lastName"`
-	Description  string    `json:"description"`
-	CreatedAt    time.Time `json:"createdAt"`
-	LastUpdateAt time.Time `json:"lastUpdateAt"`
-	OptIn        bool      `json:"optIn"`
+	PrimaryKey   string     `json:"primaryKey"`
+	SortKey      string     `json:"sortKey"`
+	Phone        string     `json:"phone"`
+	Email        string     `json:"email"`
+	FirstName    string     `json:"firstName"`
+	LastName     string     `json:"lastName"`
+	Description  string     `json:"description"`
+	OptIn        bool       `json:"optIn"`
+	LastSeen     *time.Time `json:"lastSeen"`
+	CreatedAt    time.Time  `json:"createdAt"`
+	LastUpdateAt time.Time  `json:"lastUpdateAt"`
 }
 
-func NewClientItem(creator, phone, email, firstName, lastName, description string) *ClientItem {
+func NewClientItem(creator, phone, email, firstName, lastName, description string, lastSeen *time.Time) *ClientItem {
 	return &ClientItem{
 		PrimaryKey:  creator,
 		SortKey:     uuid.New().String(),
@@ -28,7 +29,8 @@ func NewClientItem(creator, phone, email, firstName, lastName, description strin
 		FirstName:   firstName,
 		LastName:    lastName,
 		Description: description,
-		CreatedAt:   time.Now(),
+		CreatedAt:   time.Now().UTC(),
 		OptIn:       true,
+		LastSeen:    lastSeen,
 	}
 }
